@mtfishman 2024-02-21
Looks good, thanks. Could you add a test for `zero(::Type{QNVal})`?