@corbett5 2024-02-22
> Looks good, thanks. Could you add a test for `zero(::Type{QNVal})`?

Done, I added a test for `zero(qv::QNVal)` since I didn't see one for that either.